perf: fix potential stack overflow
In the tuple bench there were a couple of places which were going
to cause stack overflow, when their code was executed in a regular
context, not at the singleton construction phase.

Lets make them use the heap.

In scope of #9965

NO_DOC=internal
NO_CHANGELOG=internal
NO_TEST=it is a test