box.sequence: sequence_data recovery after reset
There are two sources of sequence state in tarantool: `_sequence_data`
table and it's light version used in `box/sequence.c` which is stored
in a hashmap. The reason for this decision was lack of cross-engine
transactions. The first is used in lua API, the second -- internally
to process memtx index updates. Now lua `reset()` call always produce
WAL entry and we don't rely on memtx tables content during the initial
recovery since the latest hasmap state is stored into snapshot directly.

Fixes #9871
NO_DOC=bugfix