box.sequence: sequence_data recovery after reset
There are two sources of sequence state in tarantool: `_sequence_data`
table and its light version used in `box/sequence.c` which is stored
in a hashmap. The reason for this decision was lack of cross-engine
transactions. The first is used in lua API, the second -- internally
to process memtx index updates. Now lua `reset()` call is always replayed
during the final recovery on one hand and we don't rely on the spaces
(other than system ones) content during the initial recovery since the
latest hasmap state is stored into the snapshot directly on the other
hand.

Fixes #9871
NO_DOC=bugfix