feature: add validate method to tuple format objects
Added a `validate` method to `box.tuple.format` objects to
simplify tuple validation against a format without
requiring tuple creation or insertion.

Closes #9979

@TarantoolBot document
Title: Tuple and space format validate
Since: 3.4

To simplify validation of tuples without needing to create or insert them,
a new `validate` method was added to `tuple.format` objects.

This method verifies that all fields in the tuple adhere to the type and
structure requirements specified in the format. If the tuple is invalid,
an error is raised.

Example:
```lua
format = box.tuple.format.new({
    {name = 'id', type = 'unsigned'},
    {name = 'name', type = 'string'},
})
format:validate({1, 'A'}) -- ok
format:validate({true, 'A'}) -- error
```