gitignore: add DartConfiguration.tcl
Seems like this file is generated in the build directory for the ancient
DART testing tool. When the build directory is the same as the source
directory, git shows it as a new untracked file. That's irritating.

NO_DOC=no code changes
NO_CHANGELOG=see NO_DOC
NO_TEST=see NO_DOC