log: fix message being truncated to 1024 bytes
Before this patch, when log format was set to 'json' and log entry was
written with format-string API (`require('log').info(<format string>,
args...)`), in the output the message was truncated to 1024 bytes.

This patch fixes this behavior, and now the message will only be
truncated if the full log entry (with all the context) doesn't fit into
the 16KiB buffer. The message will be truncated in a way, so all the
context is present.

Closes #10918

NO_DOC=bugfix

(cherry picked from commit 6b891b4f7648113ed6501d8957f07411d1a33368)