test: stabilize the gh_8802_box_session_push_deprecation test
The test assumes that as soon as one process has finished writing a log
message the other process will be able to read this message from the log
file. This would be true if tarantool process wrote logs directly to a
file, but in luatest tests the logs are passed through tee process,
which duplicates the logs to the unified log file, so
a line written to log by tarantool server is not guaranteed to appear in
the log file read by the test immediately. Add retries for the log
reading assertions.

NO_CHANGELOG=flaky test
NO_DOC=flaky test