ci: add alternative testing workflow (justtest)
The idea is to try to find a proper balance between a simplicity of the
CI script and an amount of time a developer waits for feedback from CI.
Hopefully, it is what the development team is able to maintain and
adjust for their needs on demand.

At first, this workflow doesn't use an intermediate script like
`.test.mk` and directly runs necessary commands. Also, it has no steps
defined outside (like separate actions or reusable workflows), so all
the logic is there. Only well-known actions with very clear meaning are
used here: these are `actions/checkout` and `actions/cache`.

Next, the workflow uses the GitHub Actions runners infrastructure
provided by GitHub itself. This way we don't need hacks for
always-online runners (cleaning directories, killing processes and so
on). Also, we don't have to solve infrastructural problems.

This workflow doesn't attempt to perform various things that give a
little gain for testing, but complicates the script. It has no any
secrets and don't perform deployments or instant messaging
notifications. It doesn't attempt to react on labeling the pull request
or cancel unneeded jobs based on some rules.

At the same time, in order to achieve a good timing, the workflow
actively uses caching and performs splitting of building and testing
jobs. This way it takes about 6 minutes.

The workflow file is commented with motivation behind these choices.
Check it out for details.

NO_DOC=development tooling
NO_CHANGELOG=same
NO_TEST=same