tuple: fix confusing hashing of 'number' index
Hash index over a number field was showing a confusing behaviour:
same non-integer number stored as MP_FLOAT and MP_DOUBLE had
different hashes. For example, a user could insert float 14.5 and
would be unable to get the tuple by double 14.5 key.

At the same time if the index was 'double', the scenario above
worked fine. Also it works fine in a tree index. Which means the
'number' index behaviour was not only confusing but also
inconsistent.

Lets fix it and hash all non-integer values as doubles.

Note that the hashing function uses mp_store_u64() instead of
mp_encode_double(). This is because all the doubles have the same
header MP_DOUBLE. It is pointless to include it into the hash
function. Besides, as the next commit will show, hash from an
msgpack number without a header is quite faster to calculate.

Part of #9965

NO_DOC=bugfix
NO_CHANGELOG=later