connpool: add internal pool class
This patch introduces an internal pool abstraction in
`experimental.connpool` module. The idea is quite similar to a cartridge
pool module [1] used in RPC [2]. We need to have a way to work with a
set of connections. But instead of URIs we use instance names now.

This abstraction is useful for #10330. The issue is about making the
connpool check the connections in the background. It includes closing
unused connections automatically to free a few file descriptors. We want
to make `filter()` and `call()` methods open connections that'd be
closed if they remain unused. These functions use a public `connect()`
method internally. Thereby straightforward implementation requires
extending public `connect()` method with an option making the connection
non-persistent.

This patch moves the basic `connect()` logic into an
internal pool function allowing to keep the "non-persistent"
completely internal.

This patch only involves the most straightforward implementation
of the pool as an key-value connection map with a `connect()` method
which is likely to be extended in the future. The `connect()` method
behaves the same way as it's been implemented in a public API except
instead of throwing an exception in case of an error it returns nil as
the first return value and an error message as the second return value.

Also, the pool introduces a `connect_to_multiple()` call providing the
same behavior as `connect_to_candidates()` i.e. connect to all of the
candidates and wait until the timeout or until all of them is reached,
an `any` option can be supplied to make the method wait for at least one
connection satisfying the constraints.

Needed for #10330

[^1] https://github.com/tarantool/cartridge/blob/master/cartridge/pool.lua
[^2] https://github.com/tarantool/cartridge/blob/master/cartridge/rpc.lua

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring