connpool: add internal connect method
This patch introduces an internal `connect_to_instance` method that is
intended to be used internally instead of using public `connect()` call.

The method itself starts connecting to the instance and returns
connection handle.

Technically, it's just a little refactoring for a future improvements
caused by the need to extend the internal logic with new handles and
not to touch the public method at the same time.

Also, the patch makes a small change and makes the code in `filter()`
reuse `connect_in_background()` call.

Needed for #10330

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring