connpool: add internal pool class
This patch introduces an internal pool abstraction in
`experimental.connpool` module. The idea is quite similar to a cartridge
pool module [1] used in RPC [2]. We need to have a way to work with a
set of connections. But instead of URIs we use instance names now.

This abstraction is useful for #10330. The issue is about making the
connpool check the connections in the background. It includes closing
unused connections automatically to free a few file descriptors. We want
to make `filter()` and `call()` methods open connections that'd be
closed if they remain unused. These functions use a public `connect()`
method internally. Thereby straightforward implementation requires
extending public `connect()` method with an option making the connection
non-persistent. This patch moves the basic `connect()` logic into an
internal pool function allowing to keep the "non-persistent"
completely internal.

This patch only involves the most straightforward implementation
of the pool as an key-value connection map with a `connect()` method
which is likely to be extended in the future. The `connect()` method
behaves the same way as it's been implemented in a public API.

Also, the patch makes `filter()` reuse `connect_in_background()` call.

Needed for #10330

[^1] https://github.com/tarantool/cartridge/blob/master/cartridge/pool.lua
[^2] https://github.com/tarantool/cartridge/blob/master/cartridge/rpc.lua

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring