box: set TXN_WAIT_SYNC with xrow rather than limbo state during recovery
During recovery we set the `TXN_WAIT_SYNC` flag based on the limbo's state
rather than respecting the xrow's flags. This does not play well with fully
local transactions that were committed using the `TXN_FORCE_ASYNC` flag:
during runtime they bypass the limbo as expected, but during recovery we do
not set this flag, and they are appended to the limbo, if it is not empty,
even though they are not supposed to be, since their xrows does not have
the `wait_sync` flag.

To fix this, during recovery let's set all the required flags related to
synchronous replication while processing the xrow, and ignore the limbo
emptiness, if we are in recovery and processing local transactions (i.e.,
not applying remote transactions).

Closes #11053

NO_CHANGELOG=<this version is not affected by backported bugfix>
NO_DOC=<this version is not affected by backported bugfix>
NO_TEST=<this version is not affected by backported bugfix>

(cherry picked from commit 4ba187909566e91d2a54499c2985bd2be8be1b2a)