box: set TXN_WAIT_SYNC with xrow rather than limbo state during recovery
During recovery we the `TXN_WAIT_SYNC` flag based on the limbo's state
rather than respecting the xrow's flags. This does not play well with fully
local transactions that were committed using the `TXN_FORCE_ASYNC` flag:
during runtime they bypass the limbo as expected, but during recovery we do
not set this flag, and they are appended to the limbo, if it is not empty,
even though they are not supposed to be, since their xrows does not have
the `wait_sync` flag.

While currently this only affects writes to the `_gc_consumers` space, in
the scope of #10460 we will also need to enable fully local transactions
from the `box.ctl.on_replication_split_brain_rollback` trigger to be
committed bypassing the limbo.

To fix this, during recovery let's set all the required flags related to
synchronous replication while processing the xrow, and ignore the limbo
emptiness, if we are in recovery and processing local transactions (i.e.,
not applying remote transactions).

Closes #11053
Needed for #10460

NO_DOC=<bugfix>