vinyl: use txn::psn instead of vy_tx::psn
There's no need in a separate PSN counter anymore now, when we have
the PSN counter used by the memtx engine. It's behavior is equivalent
to the counter used by vinyl. Using the same PSM counter in vinyl
and memtx is important for sending a cross-engine transaction to
a consistent read view.

Needed for #1803

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring