cmake: support ASAN and LSAN by CTest
Running tests with enabled ASAN and LSAN requires defining a
number of environment variables (like a path to a file with
suppressions or custom sanitizers flags). These environment
variables are defined in .test.mk and therefore it is complicated
to use these variables with CTest. The patch moves environment
variables to `cmake/profile.cmake` and sets them in targets with
test-run.py and tests for CTest.

NO_CHANGELOG=testing
NO_DOC=testing
NO_DEPS_UPDATE