box: make 'double' an alias for 'number'
When 'double' field type was introduced, it would only allow
MP_DOUBLE MessagePack values. This would work fine. The numbers
were compared correctly, could be promoted to 'number' and
'scalar' types.

Then the users were complaining that it is too hard to insert
integer values into a 'double' column because, at least when come
from the Lua land, the numbers with 0 fractional part are saved as
MP_INT/MP_UINT.

The commit 51af059c10ffa451aaaa59053014f11cc64f8d5c ("box: compare
and hash msgpack value of double key field as double") allowed
the users to insert MP_INT and MP_UINT into the 'double' field,
but it kept the C-double comparison logic. Meaning that all the
numbers would be cast to C-language doubles and compared like
that, regardless of their MessagePack type. It would behave like
if the numbers are all actually stored as MP_DOUBLE.

This eventually led to problems, because the assumption of
'double' being included into 'scalar' and 'number' broke. The same
values stored in a 'double' index and in a 'number' index would be
sorted differently, because large integers, when cast to C-double,
loose precision. In double-world UINT64_MAX == UINT64_MAX + 1,
while it is obviously not so in number- and scalar-world.

The index simply broke when it was altered to 'number' or
'scalar', because it doesn't get rebuilt, and with the new sorting
rules it became sorted incorrectly. This led to all sorts of data
inconsistencies, assertions and crashes.

With vinyl it was particularly bad, because it can't even be fixed
automatically. If a user ever had a vinyl 'double' index with
large integers and then converted it to 'number' or 'scalar', then
the index is broken. And it is undetectable from just looking at
the field types. The unlucky users of vinyl with 'double'-type
indexes are going to have to pay attention to the next release and
rebuild their indexes manually.

Closes #9965
Follow-up for #7483

@TarantoolBot document
Title: `double` field type new behaviour

`Double` field type now is identical to `number`. I.e. it can
store integers and compares them as integers.

* [Reference -> SQL reference -> SQL user guide](https://www.tarantool.io/en/doc/latest/reference/reference_sql/sql_user_guide/#sql-data-type-double):
*DOUBLE values are numerics that do contain decimal points (for
example 0.5) or are expressed with exponential notation (for
example 5E-1)* -> now DOUBLE can have any numbers, including
integers. It becomes identical to NUMBER.

* [Platform -> Defining and manipulating data -> Data storage](https://www.tarantool.io/en/doc/latest/platform/ddl_dml/value_store/):
*the storage type is MP_DOUBLE and the size of the encoded
value is always 9 bytes. In Lua, fields of the double type
can only contain non-integer numeric values and cdata values
with double floating-point numbers* -> not so anymore.
`double` field is now identical to `number` (it can store
integers too, and compares them like integers).

(cherry picked from commit 60272e8292ad5a893cc14d0318a30f5cfabb6495)