replication: fix user notification to leaving orphan mode
The problem is that a call to `box_do_set_orphan(true)` can silently
change the value of `orphan` to true without notifying the user, and a
subsequent call to `box_set_orphan(false)` will cause the user to be
notified to leaving orphan mode. This may confuse the user.

This patch notifies the user to leaving orphan only if there was
previously a notification to enter.

Part of #10592

NO_DOC=bugfix
NO_CHANGELOG=Will be added in another commit
NO_TEST=Will be added soon