replication: go to 'orphan' before synchronizing with replicaset
There was a bug when a master who crashed and lost his xlogs could not get
his own rows when it tried to re-join the cluster and hangs forever.
The bug was due to the fact that at this moment the master is not in the
"orphan" state (but it should be in it, because it is not actually
synchronized with the quorum), and therefore believes that it can filter
its own records.

Only with `bootstrap_strategy=='LEGACY'` it was not reproduced. This was
due to the node silently going into an orphan state when it had to
reconnect to the replicaset for some reason.

The patch solves this problem in the following way - tarantool silently
goes into the "orphan" state upon any reconnection (full or partial).
It should be fair. In the case of "orphan", a false positive is better
than a false negative. The main thing is to ensure that if a node is not
synchronized with quorum, then the node is in an "orphan" state. It will
not scare users in any way, because they will only be notified about this
when an attempt to synchronize actually ends in failure.

Closes #10592

NO_DOC=bugfix