box: introduce "waiting_for_own_rows" mode
Before this, there was a special mechanism that determined whether an
instance had received all of its rows that it lost during a restart -
to do this, it simply checked that it was currently in the orphan status.
This approach can be called a heuristic, which, unfortunately, does not
work in specific cases.

Instead, it would be much more "correct" for an instance to look at the
maximum lsn among the replicas in the replica set for the component
corresponding to that instance. This will work in all the same cases
where the previous approach worked, but in many more. But it is worth
noting that this approach does not guarantee that one day an instance
will not unexpectedly receive some of its transactions from the past
that have already been overwritten by a new transaction.

Now if an instance realizes that it hasn't received all of its
transactions yet, it goes into "waiting_for_own_rows" mode. This mode
is another ro-reason similar to "orphan" mode. While the instance is in
"waiting_for_own_rows" mode it is read-only.

Closes #10592

@TarantoolBot document
Title: new `box.info.status` - "waiting_for_own_rows"
Product: Tarantool

Root document: https://www.tarantool.io/ru/doc/latest/reference/reference_lua/box_info/status/

The `waiting_for_own_rows` status can be relevant if a node that was once
a master crashed and lost some of its xlogs for some reason (e.g. manually
deleted by the user). In this case after rebooting, `box.info.status` will
be equal to `waiting_for_own_rows`, which means that the instance knows
that some of its transactions were lost from its xlogs and it is waiting
for them to be received via replication. While instance status is
`waiting_for_own_rows` it remains read-only.

(cherry picked from commit 61cb5b5e3414a2ac0aac5682bfd2435457ae87bd)