memtx: do not use ER_CROSS_ENGINE_TRANSACTION for alien snapshot row
If a memtx snapshot contains a row for a different space engine,
ER_CROSS_ENGINE_TRANSACTION is raised. This is confusing because
this errors is normally raised on an attempt to start a cross-engine
transaction. Besides, when support for cross-engine transactions is
finally added, we'll want to remove this error. So let's introduce
a new error code for this case and add a test.

Needed for #1803

NO_DOC=minor
NO_CHANGELOG=minor