config: allow failover only in the global scope
This patch adds validation on the scope of the `failover` section in the
instance configuration. This section configures a EE-only supervised
failover and it should be configured only in the global configuration
scope. This patch implements this restriction.

NO_DOC=it's already marked as global only in docs

(cherry picked from commit f63a8dad657a66db51dc44eb3b0318cf90dd1d89)