netbox: fix in-progress request counter underflow during schema fetch
If the remote schema changes while the client is fetching it, it will
restart the procedure. This effectively abandons the requests issued
in the previous attempt to fetch the schema. The server will still reply
to them though - when this happens, the client will subtract the counter
of in-progress requests, underflowing it, because these requests were
never accounted in the first place.

In case of a debug build, this bug results in a client crash due to an
assertion failure. In case of a release build, the client doesn't crash,
but the server will hang on shutdown because the counter is used for
implementing the graceful shutdown protocol.

Let's fix this issue by accounting these abandoned requests when
we restart fetching the remote schema.

Closes #11062

NO_DOC=bug fix

(cherry picked from commit 2df320405d21cea46f76bba5a686bfd848c8b278)