memtx: do not run MVCC GC in handlers called by read requests
Current design of memtx MVCC garbage collector produces to major
problems. Both problems are caused by implicit index modification - when
MVCC deletes the last story of an actually deleted tuple, it has to
remove the tuple from the index.

The first problem is if there is a functional index in the space, replace
will call a Lua function to obtain the key. However, read methods in memtx
(e.g. `select`) are called with FFI - it is faster than Lua/C, but have
some restrictions, and one of them - Lua call during another FFI call is
forbidden. So, using MVCC with function indexes can easily lead to a
crash.

The second problem is that we have to carefully use memtx handlers
because they can modify index and break its iterators, so we always
write a large "MEMTX MVCC GC BOUND" notification next to every such call
and sometimes manually call the garbage collector.

The solution of both problems is simple - just do not call GC in
handlers called on read requests! But won't it slow down the garbage
collector? Actually, almost nothing will be changed. That's how GC works
now - on every creation of a memtx story it accumulates amount of steps
for the story GC. Then, later, when the GC is called, it spends all the
accumulated steps. The stories are created only in two cases:
1. Insert of a new tuple. Story is always created.
2. Read of an existent non-dirty tuple. Story is created only if there
   is an active transaction (IOW, only in non-autocommit mode).
The commit removes calls of GC in handlers used by read requests, so now
it is called when:
1. A write (insert, delete or update) happens.
2. Statements get prepared.
3. Statements get committed.
4. Helper `memtx_tx_clean_txn` is called - actually, it is called for
   each transaction when it is destroyed.
So, after each story creation there will be call of the GC, hence it's
useless to call the GC on read requests. Moreover, it leads to lots of
problems. Let's just not do this.

Closes #7867
Closes #10770
Closes #10775

NO_DOC=bugfix