test: stabilize engine/errinj_ddl test
The test relied on fibers finishing execution in particular order and
setting/unsetting error injections in time. Of course it was flaky. In
all the 3 flaky cases the fiber unsetting the injection could be
executed before the fiber which sets the injection, and the rest of the
test would run with that injection being set.

Fix this properly and remove the test from the fragile list.

Closes tarantool/tarantool-qa#20

NO_CHANGELOG=flaky test
NO_DOC=flaky test