memtx: set result on deletion from functional tree index
All indexes return deleted tuples as a result (in memtx, at least), but
by mistake we forgot to set result in functional memtx tree index - the
patch fixes the mistake.

Note that this small inaccuracy was breaking memtx MVCC since we rely on
the replace result there.

NO_DOC=bigfix