replication: fix calculation of sync quorum in 'config'/'supervised'
There was a bug related to synchronization with the replica set in the
'config' and 'supervised' bootstrap modes: the node didn't go into an
'orphan' status even if it failed to synchronize with each of the
connected nodes.

The bug was related to how the node calculated synchronization quorum in
the 'config' and 'supervised' modes of bootstrap. Because of this, in
many cases the quorum turned out to be 0. This meant that the node would
not go into 'orphan' even if it was not synchronized with any of the nodes
specified in `box.cfg.replication`.

This commit fixes this bug.

Closes #11156

NO_DOC=bugfix
NO_TEST=TODO