txn: do not rollback transaction statements on timeout
The transaction timeout may be triggered by vinyl taking too long to
read disk while executing a statement. If we roll back the transaction
in this case, the statement will be deleted, resulting in a crash when
we finally try to commit the statement. Actually, in contrast to the
case when a transaction is aborted by a fiber yield, there's no need to
rollback the transaction on timeout  We don't rollback the transaction
on conflict so let's not do it on timeout, either.

Closes #11088

NO_DOC=bug fix

(cherry picked from commit a4b9e5dc3f0061cccb657e4935a09228bc61b31d)