replication: refactor `replicaset_connect` flags
`connect_quorum` was intended to have the following semantics:
`true`	- crash if the node couldn’t connect to the quorum.
`false`	- go to orphan before reconnecting, and stay in it if the node
couldn’t connect to the quorum.
But there are additional confusing properties that should not logically
depend on `connect_quorum`: if `connect_quorum` is `false`, it does not
wait for all connections, only for a quorum. But this is not logically
related to how we want to handle quorum connection failure - to fail or
to be in "orphan". It seems to me better to introduce a separate argument
that will determine this.

Now instead of the `connect_quorum` argument there are two others:
`wait_all` - whether replicaset_connect should wait for all connections
to be established for a full `replication_connect_timeout`.
`demand_quorum` - whether `replicaset_connect` should fail in case it
can't connect to a quorum.

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring