test: stabilize replication/sync test
When replication is restarted with the same replica, there's a chance
that an old relay may be not stopped yet when a new applier connects.
In this case the applier should get ER_CFG error and it should continue
trying to sync with master.

The test stopped the relay from exiting with slowing sending of the rows.
This is not reliable and sometimes relay successfully exited before
replica tried to connect, which leads to the absence of the ER_CFG
error. Let's block relay exit with ERRINJ_RELAY_EXIT_DELAY instead,
which is much more reliable.

Closes tarantool/tarantool-qa#6
Closes #4129

NO_CHANGELOG=flaky test
NO_DOC=flaky test