box: introduce replication_reconnect_timeout
Previously `replication_timeout` was used as interval to
sleep between applier reconnects. However, `replication_timeout`
shows, how often heartbeats are sent, and has nothing to do
with reconnect timeout.

The problem with using `replication_timeout` as reconnect
timeout is that it's difficult to set big values to
`replication_timeout`:

  1. Some instance gets configuration change {replication_timeout =
     10000}. Other ones haven't got it yet. The relays of that
     instance stop sending hearbeats. Appliers on the other side
     exits with TimeOut error and try to reconnect every
     `replication_timeout`.

  2. At some point of time new configuration comes to other instance.
     It may turn out that `replication_timeout` change comes during
     failing reconnect, which causes 10000 hang before the next
     reconnect and replication cannot be restored without restarting
     the replication (or waiting for timeout).

In order not to introduce compat, new option by default equals to
`replication_timeout` (it's set to 0).

Closes #11115
Part of #10059
Needed for tarantool/tarantool-qa#333

@TarantoolBot document
Title: `replication.reconnect_timeout` option

Since 3.4.0. The timeout (in seconds) between attempts to reconnect
to a master in case of connection failure. Default is 0. If the
option is omitted or set to 0, then it equals to reconnect_timeout.